adds a plain old webpack-cli task
