+++ packages/just-scripts/src/tasks/index.ts
@@ -12,4 +12,5 @@ export * from './upgradeStackTask';
 export * from './copyInstructionsTask';
 export * from './prettierTask';
 export * from './eslintTask';
+export * from './webpackCliInitTask';
 export * from './webpackCliTask';
